refactor: move test pacakge to a unique path into column test
