+++ jnosql-lite/mapping-lite-column-test/src/test/java/org/eclipse/jnosql/lite/mapping/entities/ColumnEntityConverterInheritanceTest.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.lite.mapping.entities.column;
+package org.eclipse.jnosql.lite.mapping.entities;
 
 import jakarta.data.exceptions.MappingException;
 import jakarta.inject.Inject;
