+++ jnosql-lite/mapping-lite-document-test/src/test/java/org/eclipse/jnosql/lite/mapping/entities/PersonRepositoryTest.java
@@ -279,7 +279,6 @@ class PersonRepositoryTest {
         });
 
     }
-
     @Test
     public void shouldFindPageableLimit(){
         when(template.select(any(DocumentQuery.class))).thenReturn( Stream.of(new Person(), new Person()));
